Add options to reduce size and fit 64KB devices again
This adds 3 new options to configure:

  --disable-flash-upgrade: To disable special commands
    that enable flash programming via regnual.
  --disable-rsa-support: Disables RSA algorithms, which
    leaves only ECC as usable. Can't be used with the
    enable-flash-upgrade knob since it requires RSA.
  --slow-crypto: Reduces size of SHA functions at a
    cost of making it slower (~15KB saved).

Modified the --with-dfu option to take an argument that represents
the DFU stage size in KBs.

It also replaces polarssl AES with a tiny AES implementation
that saves space by not using 8KB of lookup tables.
Reduced the first 4K pages to just 1KB since we do not have AES
tables anymore. Also removed some minor alignments in the linker
script to save more space.

Disabling RSA and optimizing for size results in 58KB binary,
which is able to accomodate a 6KB DFU in a 64KB device.